# openapi.yaml
openapi: 3.0.0
info:
  title: API LMS - Plateforme de gestion d'école
  description: "Examen Mi-Semestre L3 Informatique - Système complet de gestion scolaire."
  version: 1.0.0
  contact:
    email: admin@ecole.com
servers:
  - url: http://localhost:8000
    description: Serveur Local de Développement
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Student:
      type: object
      required: [name, email]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Jean Dupont"
        email:
          type: string
          format: email
          example: "jean@exemple.com"
        phone:
          type: string
          example: "0340000000"
        enrolled_at:
          type: string
          example: "2026-05-14 10:30:00"
          description: "Format requis : YYYY-MM-DD HH:MM:SS"
...
paths:
...
  /api/v1/students:
    get:
      security: [{ bearerAuth: [] }]
      tags: [Students]
      summary: Liste des étudiants
      parameters:
        - name: name
          in: query
          required: false
          schema: { type: string }
      responses:
        200: { description: OK }
    post:
      security: [{ bearerAuth: [] }]
      tags: [Students]
      summary: Créer un étudiant
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Student' }
      responses:
        201: { description: Créé }
  /api/v1/students/{id}:
    get:
      security: [{ bearerAuth: [] }]
      tags: [Students]
      summary: Détail étudiant + enrollments
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        200: { description: OK }
    put:
      security: [{ bearerAuth: [] }]
      tags: [Students]
      summary: Modifier un étudiant
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Student' }
      responses:
        200: { description: OK }
    delete:
      security: [{ bearerAuth: [] }]
      tags: [Students]
      summary: Supprimer un étudiant
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        204: { description: Supprimé }

  # --- COURSES ---
...
  /api/v1/webhooks/payment:
    post:
      tags: [Webhooks]
      summary: Webhook de paiement
      parameters:
        - name: X-Webhook-Signature
          in: header
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                event: { type: string, example: "payment.succeeded" }
                data:
                  type: object
                  properties:
                    enrollment_id: { type: integer, example: 3 }
                    amount: { type: integer, example: 5000 }
      responses:
        200: { description: OK }